Rename Legionella to Geobacter.
When I originally picked the name Legionella, I was unaware that
Legionella the genus can live in Humans, and have in fact killed
many! Which is sad, and makes it not a great thing to use as a name.

Geobacter, OTOH, was the first discovered bacteria to essentially
feed on rust. Which I think I like better anyway.